Merge venue and location columns
Some parks and park systems are so large as to functionally be
multiple orienteering venues. The intention of venue+location was
to be a unique identifer for each orienteering venue, but location
names were inconsistent, and sometimes captured meaningless
differences (i.e., Shelter 3 vs Shelter 4). The new venue name column
now has a one-to-one mapping to maps. A different venue name means
a different map.